[sherlock][ge2d] Fix GE2D protocol to add callback for SetOutputResolution.
Use the existing ImageFormat_2 structure in Init to set the initial
resolution. The drivers make a copy of the table of Image Formats. For
any subsequent SetOutputResolution calls, the caller passes the index of
the new ImageFormat to use. SetOutputResolution needs to be an Async
operation, notifying the caller after the necessary sub-tasks for changing
the resolution are complete via a callback (which also contains the
index of the new output resolution).

Tests: No new tests are needed, since this changes the GE2D protocol. I
ran the GDC task tests to ensure there are no regressions.

Change-Id: Idebf0b05cdcdd9d361b93644e6769d0e107fd51a