[fs-management] add configuration options
There are various options that can be passed to the underlying
filesystem process. This adds the ability to configure which options to
pass. Unlike the c fs-management library, the options are set on the
filesystem manager for all calls that go through it, instead of at the
call site for the operation, and only options that configure the
underlying process are included.

Test: fx run-test fs-management-test
Bug: 35774
Change-Id: I315de00701897c803e8cef62751d7d6eee66d9de