[fs] Make ManagedVfs thread-safe
The current implementation requires that all interaction with a
ManagedVfs take place on the same thread as the ManagedVfs's async loop.
The most straightforward way to use a ManagedVfs in tests is to spawn a
separate thread for the Vfs, and to let the main test thread run the
test logic itself.  However, this was unsafe, and that unsafety has
caused assorted crashes and flakes.

To avoid folks having to rediscover this again, let's just make
ManagedVfs safe to access from multiple threads, as the underlying Vfs
is, by guarding access to the interesting members with a mutex.

Bug: 36485
Change-Id: I1d1c05147fcfdfcfa685b940808f5e164ba98b82