[bt][lib] Adapt fuchsia_bluetooth::types:Uuid to be backed by the uuid crate
The Uuid type in fuchsia_bluetooth::types is now a newtype wrapper
around the `Uuid` exposed by the third party rust `uuid` crate. This
allows us to use their existing parser for uuids.

Added property-based tests to verify the roundtrip validity of the
parser, and that the .unwrap() in .from_bytes() is safe in all cases.

test: fx run-test bluetooth-crate-tests

Change-Id: Id97e76ba9a8c61668b78387e924c441da8005a17