[fidlcat][fidl_rt] Split FIDL decoder into its own lib
Fidlcat currently has the only FIDL decoder that can marshall at
runtime. We're going to need this for shell work, so let's split it out.

This is a fairly straightforward split. It may over- or under-fit by a
bit, but it's the simplest split we can do, and we can move smaller
pieces back and forth later.

Change-Id: I1be4cab8ebaaa0460ba7eb15ddff493521a5cec1