[system-ota-tests] Support longevity testing
This refactors the OTA test to support longevity testing, where a
device continuously updates to the latest available version, or errs
out if a problem occurs. Since the test now has two modes, it removes
the notion of running subtests, which is a little complicated to mix
in. Instead, it will now skip doing the downgrade OTA/downpave if no
pavers/downgrade builders/etc are not specified.

Along the way, it also return better error messages if a
builder/build id/repo is not specified.

In addition, this switches the upgrade_test to use `PATH= rm` to trigger
garbage collection in order to use the command that is baked into dash,
rather than `rmdir`, which is a separate command and might not be a part
of the base package.

Bug: 35885

Change-Id: Ic373553dac24d7f3ab8c9219feb7110ab2a51f51