[pseudo-fs-mt] {Token,Inode}Registry
Preparation for the client side mutable directories.  We would need
registries that would allow directories to exchange tokens for the
`Rename` and `Link` operations.

While at it, also add an `InodeRegistry` that allows entries in the tree
(both directories and non-directories) to acquire unique "inode" ids to
be reported to the clients.  This will allow users of the library to
address fxbug.dev/4630 for pseudo-fs-mt.

Both registires will be scoped using `ExecutionScope` to cover more use
cases for the users of the library.

Bug: 4630

Change-Id: I2e5d302efa9d4c7b19e6028d3ab9662af15e9b57