[devcoord] More SystemInstance refactoring
This is the third patch in a series of patches which all make no
functional changes to the devmgr suite, but refactor things to untangle
some global dependencies and static functions.

Now that we've extracted SystemInstance into its own source file, start
making fields private, where appropriate.  Order the methods in
system-instance.h by the order in which they are called in normal
startup.

Change-Id: Ie4f917549de16ccf09c9f460eaea3c2d0928805a