[Zircon] Cleanup instances of `-Wsizeof-array-div` warnings
Clang introduced a new warning that is raised when we
attempt to check the size of an array through sizeof() division. If the
type of the sizeof() in the divisor is not the same as that of the array
element, a warning is raised.

Although there are no functional changes in this patch, this keeps the
fuchsia.clang bots green.

BUG: 36439
Change-Id: I3da6eb6c6987b5a161e4df4402d9f832860fac89