[fshost] Avoid checking minfs devices without a DATA GUID
This patch is useful for tests of external partitions containing
minfs devices. Without this patch, those minfs partitions will
be "checked / auto-formatted on corruption" at boot.

With this patch, this will only happen for a minfs patch which
has the accompanying type GUID. This provides parity for the mounting
decisions made for blobfs.

TEST: block-watcher-test expanded

Change-Id: I24291b0a1ef7e41a7ea96aeaef22a027ab3d3aa1