[devmgr] extract DevmgrLauncher, FsProvider
This is the fifth patch in a series of patches which all make no
functional changes to the devmgr suite, but refactor things to untangle
some global dependencies and static functions.

This removes the last static globals from devcoordinator.
`fs_clone` is no longer bound to executables by the linker, but
explicitly provided to the DevmgrLauncher instance.  As a result, some
of SystemInstance's previously public fields can be made private, as
the CloneFs implementation moves into SystemInstance.

Change-Id: Ie225bf3dd555694abebad42440f3450303a39048