[componentmgr] Added a test for visibility of eager children in Hub
Added a test that verfies that grandchildren are not shown in the hub
when the children are lazy.

To do this, we set up the following component tree and query the children
of B via the hub. C should not be shown as a child of B.

A -> Parent (started by test)
|
B -> Child (lazy)
|
C -> Grandchild (eager)

Test: Integration test written to hub_integration_test.rs
Change-Id: Ie39aaeacbee9d941bf91db7d24db1815742e9784