[thermd] Create separate manifests for x64 and arm
The thermd service uses root resource to issue the power syscall on Eve.
For Astro, the root resource is not needed. The proer way to fix this is
to write a board driver that gets root resource, issue the syscall, and exposes
a fidl api for services like thermd. This is a termporary workaround to
avoid the security issues on Astro.

BUG: fxb/34164

Test: fx build. On astro, verified that the thermd is started and trip
point updated.

fx set core.astro --release --with-base ///garnet/bin/thermd:thermd_tests_package
fx run-test thermd_tests_package

Change-Id: Iee3ff3675a014c7c824ec109ae08238675938270