[componentmgr] Hooks event propagation
This CL implements hooks event propagation.

A couple of changes are made here.

1. Hooks was moved from Model to Realm.
2. There is a parent / child relationship between Hooks objects.
   An event in the child propagates up to the parent.

If hooks are installed in a parent realm, then all events within that
containing realm will bubble up to that parent and any hooks installed
on that parent will be called out to.

This CL enables scoping of hooks in the future.

Change-Id: Ia4971f7ae4b25fe465ac3e152b99d60cba2473cc