[media][tests] Phase Response tests for resamplers
This test-only CL adds phase response to the set of measurements performed
by audio_fidelity_tests. This is measured in the same analysis pass as
frequency response, SINAD and out-of-band rejection -- that is, each
sampler is tested at 8 different resampling ratios across various
frequencies (3 by default in CQ, but 48 if '--full' is specified).

This is being added in preparation for new resamplers with higher
fidelity requirements.

For our existing resamplers, note that SampleAndHold has linear
phase response wrt frequency, while Linear has zero-phase.

Test: creates 20 new test cases

Change-Id: Ie326a1b2cc17b6fd270b1221afeae8c7f25b4cb4