[debugger] Add a new index implementation.
The new index is not currently hooked up. It currently exists in
parallel to the old one and is only exercised by tests. The old system
will be switched in a followup to make the landing easier.

The first important change is in the structure of the index. In the
previous implementation everything with the same name was grouped in the
same IndexNode. This made it hard to search for things of a specific
type. The most difficult use-case is to search for a matching function
in any namespace, which requires a full brute-force search.

The new index has a separate structure for namespaces, functions, types,
and variables. This makes it more efficient to walk when looking for one
specific thing, although it makes the walk itself more complicated if
the caller doesn't know the exact type of everything. Walking the new
index is not currently hooked up.

The second change is how the index is built. The code is cleaner and
more efficient by extracting all information from the DIE in in the
first scan pass and using this in the second indexing pass. The previous
implementation extracted the names in the second pass which meant some
duplicate DIE decoding. The new implementation also caches previous
index insertions which saves some map lookups. The time to index a full
Chrome release binary dropped from 47s to 32s.

Change-Id: Ic5b335665713413633bd487a8087b2719d3c8835