[componentmgr] Fix component_manager flake
Fix a bug with a high flake rate in CQ. Why the bug is being triggered
isn't clear yet, but it's obvious the bug is there and it's likely
responsible for the flake.

The bug is a race with the realm state lock when a component instance is
started. Fix by holding the lock while starting the component.

Bug: 36469
Change-Id: Iaf940229d69d09de58f2daefff253a522d9271e7