[devcoord] Move most globals into a class instance
This is the first patch in a series of patches which all make no
functional changes to the devmgr suite, but refactor things to untangle
some global dependencies and static functions.

This has no functional change; it is purely a refactoring of the
existing code into a class to hold global state.  The next patches in
the series will extract this class into a separate file.

Change-Id: Ie817cee0809a43b989df4dbeb3d953b740fabe0f