[debugger] Add a const value to Values.
C++ constexpr values are encoded as DW_AT_const_value on either a
DW_TAG_variable or a DW_TAG_member.

This patch adds a ConstValue class to hold this optional data on the
Value base class. A test is added to ensure decoding is correct.

These constant values are currently unused in the expression system so
this patch will not change observable behavior. That will be hooked up
in a separate pass.

BUG=5723

Change-Id: I4ebff73a9cbb2bcb7b34111d8609c66ad6d33748