[debugger] Support constexpr members
Adds support for reading variables and data members declared with the
DW_AT_const_value attribute. In C++ these are generally "constexpr"
members.

Fix a bug in the implementation of unary operators where the concrete
type was not resolved, so if ! was used on a typedef or something it
would error instead of doing the operation.

Uncomment the fbl::String pretty-printer which now works (the
expressions it used were implemented with constexpr values).

Clarify documentation around FoundMember and DataMember offsets. The
const_value case is a new one where the offset is not used.

BUG=5723

Change-Id: Ibbeff6d99e89e70b6ea1dce56dd01e89053065d5