[devcoord] Order system-instance.cc to match header
This is the fourth patch in a series of patches which all make no
functional changes to the devmgr suite, but refactor things to untangle
some global dependencies and static functions.

The system-instance.h header sorts its methods in the order that the
functions are called by main.cc during normal startup.  This makes the
flow slightly more readable.

This patch contains no functional changes; just reorderings.

Change-Id: I82a4b3e74f425ea14eb2b601a7f4a44854fcdac2