[minfs] Actually replay the journal during fsck
On real devices, fshost checks the consistency of minfs
and uses that decision to decide whether or not to re-format
the underlying filesystem.

Before this patch, that decision would be reached by evaluating
the underlying filesystem without replaying the journal. This means
that although the filesystem would be safe from power failure for
"mount a minfs filesystem", it would not practically experience
that benefit on real hardware, since the device would be reformatted
before replaying the journal.

TEST= Blackout stress tests (See: tools/blackout/README.md).
Without this patch, these tests would result in filesystem
corruption after 1-5 iterations.
With this patch, these tests should be runnable indefinitely.

Change-Id: I2b89b3959e7c7e28fc145e8092a01b72d4e52118