[ledger] refactor rust cloud error handling
The cloud provider now prints nice messages when the client misbehaves.
This also clearly separate internal cloud provider errors (we crash in
those cases), client misbehavior (we return an error through FIDL and
log an explanation) and normal interactions that happen to return a
non-OK status (we translate the result of the operation to a status
just before sending it to the client, but it is otherwise not treated as
an error).

TEST=no behavior changes except for log messages

Change-Id: Ic10d9db8ab9e9cadcd647f18b5ffda349623867b