[ledger][p2p] send head on watch request
When a peer sends a watch request, send the head commit if there is
only one head. This allows the peer to catch up.
Peers receiving these heads should not add them until they have the
device marked as interested, otherwise they will not be able to request
objects from these devices. This is implemented by requiring the peer to
be marked as interested before submitting a commit batch to storage.

Bug: 36108
Change-Id: I793d9776127fe3a442bec3f540cb7678f63e8634