[Zircon][NFC] Supress warnings generated on -Wsizeof-array-div warnings
https://reviews.llvm.org/rGb8185153f35446c2a8db48ee711d2fb577674c18
introduces an update to silence some false positives for
-Wsizeof-array-div warnings. This was a followup to
https://reviews.llvm.org/D67287 which added the -Wsizeof-array-div
warning. The waring itself is meant to point out expressions that try to
calculate the number of elements in an array in the form:

```
sizeof(array) / sizeof(type)
```

where the warning is raised if the size of `type` is not the size of the
element in `array`. This warning still appaears though for instances
where such a division is intended.

The new patch by the same author added a workaround for this by just
adding parenthesis around the `sizeof(type)` which is what this CL does
to supress the warnings where behavior is intended.

This CL provides no functional changes, but will allow for the
fuchsia.clang bots to be green again.

BUG: 36439
Change-Id: Iff80f90fbb0fd425c0d62438f3e90b1c5ecb624f