[componentmgr] Add tests to verify visibility of children of dynamic children
When a dynamic child is created, their realm is not immediately resolved.
Hence, on creation, we are not aware of the children of the dynamic child.
The hub only learns about the children of the dynamic child on bind.

This change verifies this behavior.

Added a helper in hub_integration_test for verifying directory contents.

Test: Made changes to hub_integration_test and reran tests.
Change-Id: I35c65a643d280d5c776f07dc9eef77e8cc3e2e3a