[debugger] Serialized step over.
Step overs are now executed one at a time, meaning that at any given
time, if a thread is stepping over a breakpoint, it is the only thread
of that process that's not suspended.

This removes a lot of complications with races and coordination when
multiple threads hit multiple breakpoints at the same time.

This serialization is per-process, so multiple processes can be stepping
over breakpoints at the same time without interfering with each other.

Bug: 5907
Bug: 6568

TEST=Added 2 unit tests to test behaviours.
     fx run-test debug_agent_tests

Change-Id: Id0b3e554b89d44fb5a54bc50ee8781729879f91d