[settings] add capability for hanging gets to add in a change function
Change function allows us to filter out some changes to not cause a
return of the watch function to limit the rate.
For example, we can filter out changes below a certain threshold.

Refactored out the sending logic into HangingGetController to enable
better unit tests of the different possible conditions.

test: fx run-test setui_service_tests

Change-Id: I0e9a10c93b94d58f3c9c73035e749e7d016b043d