[intel-hda] Simplify Pixelbook Eve audio topology.
SST playforms such as the Pixelbook Eve require the DSP to be set up
with pipelines and modules, describing how audio should be sent from
the host to the audio codecs (and, similarly, captured again).

Our current Pixelbook Eve audio topology is as follows:

                         +--- Pipeline 1---+    +--- Pipeline 2 --+
  Playback:  Host DMA -> [Copier] -> [MixIn] -> [MixOut] -> [Copier] -> I2S

  Capture:   Host DMA <- [Copier] <- [MixOut] <- [MixIn] <- [Copier] <- I2S
                         +--- Pipeline 4---+    +--- Pipeline 3 --+

The MixIn/MixOut modules are intended for mixing together multiple
streams (MixIn), or making multiple copies of a single stream that can
be send to multiple other modules (MixOut). In our straight-line
topology, however, we don't need this functionality.

This CL simplifies the topology to just a single pipeline and two copier
modules:

                         +--- Pipeline 1 ---+
  Playback:  Host DMA -> [Copier] -> [Copier] -> I2S

  Capture:   Host DMA <- [Copier] <- [Copier] <- I2S
                         +--- Pipeline 2 ---+

A single copier can't copy from I2S to Host DMA; so we need two.

Test: No behaviour change. Existing integration tests verify that set up
of the DSP works.

Change-Id: I3a49ceb03e8e06f09add3ee58f6165d94e3a2548