[camera] Adding GetChannel2 to driver interface
As part of our migration to the fuchsia::camera2 protocols,
I am upgrading our virtual camera to support both the new and the old
protocol.  To choose which protocol the you wish to use, I am adding
an extra call to the fuchsia hardware camera protocol, GetChannel2.
If the client calls GetChannel2, they will be connected to a channel
serving the protocol defined in fuchsia::camera2::hal.

Testing: this is a FIDL only change.

Change-Id: I71ce09304068ac4a3c5ed8c5c4146789b0c43c1e