[intel-dsp] Set up topology for Pixelbook Eve headphone.
This connects the Pixelbook Eve's ALC5663 codec to the DSP, creates a
new stream, and allows playback through the device.

This CL also renames generic-sounding names such as "FMT_I2S" to their
concrete values used on Eve, such as "ALC5663" and "Max98927". The
previous values are Eve-specific, but the names gave the impression that
they were more generic than they actually are.

Later CL's _will_ need to make the topology code more generic, but in
the short term, we give structure names their real, concrete names to
avoid confusion.

Change-Id: Ic8fc8e9297a0ce8ca9faa56f15ec48d005c54a68