[alc5663] Reduce default volume of ALC5663.
We previously set the digital volume to its max value (also the hardware
default), but have observed what looks like corruption of the signal on
actual hardware. (It almost looks like the ALC5663 is overflowing,
causing peaks in the signal to suddenly jump from a high value to a low
value, and then back again.) The end result is significant distortion to
the signal.

This new value --- found by inspecting and listening to sine waves
coming out of the ALC5663 at various levels of digital volume ---
appears to avoid this distortion.

Change-Id: I73f012b34bd02c88ff8e2d30aa405e19bf5391f0