[zircon][lib][paver] Updated FVM paving logic.
An FVM is paved if the image used to pave contains a non zero
maximum_disk_size that is bigger than the maximum size exposed by the
VolumeManager(VolumeManagerInfo from VolumeManager::GetInfo).

This is in addition to a different slice size, meaning that the image is
considered incompatible if it cannot address at least the same size the
sparse_image_t claims it will eventually grow to address.

Added test for such case.

TEST=paver-test

Change-Id: I8ad6bf70a6c8de4e0da7853051102d0b0596451b