[modular] Fix flakey modular_test_harness_test
This is due to a race between two sequence of events:

1) TestHarnessLauncher asks modular_test_harness.cmx (and therefore
modular and the fake session shell) to terminate cleanly. current
thread is blocked until this happens.

2) modular dispatches Lifecycle/Terminate to fake session shell, but it
doesnt respond (because thread is blocked due to 1) above). so session
shell is force killed and OnDestroy() is dispatched *before* the
buffered Terminate() is processed.

When OnDestroy() is processed first, Terminate() is ignored and the test
hangs.

This patch spins the FakeSessionShell in a separate thread so that
Terminate() can be processed in time. Also, the test finishes 1 second
earlier!

Test: Before this fix, the test used to flake when run 100x.  Now, the test doesn't fail when run 1,000x times.

Bug: 36699
Change-Id: I6bf93358fd4a81c32538b327e3d42a4ad0f926db