[size_checker] Add -sizes-json-out flag, enhance tests
The -sizes-json-out flag will be set by the infra in order to display
differences in the sizes in the Gerrit UI.

In order to test this change, I add Test_processInput, which is
essentially an integration test for most of the code in size_checker.

To make the code more testable, I change buildDir from a global variable
to a local variable, which required passing it into functions as
an additional argument.

Unrelated minor cleanups:

* Rename blobJson to blobJSON as suggested by go vet.
* Improve error messages.
* Explicitly handle failure of ParseInt.

Bug: 36667
Change-Id: I3dc51ddbc5544b2b38ef2801760847148396f7d1