[run_test_component] Allow using vulkan and sysmem system services
The vulkan driver is system-specific (at least until we have
swiftshader), so it should be retrieved from the system
fuchsia.vulkan.loader.Loader implementation. The driver communicates
with sysmem, so it needs the system sysmem as well.

Currently apps inject the system implementations of the services, which
happens to work because there's only one implementation and it just
looks for static information from the system. In the future that may
change.

Bug: 4506

Change-Id: If2a047443966ce38f77aeb4f24bae417ad5da20b