[scripts] Teach move_source tool to make forwarding go_library targets
This teaches the move_source tool how to create forwarding targets for
Go libraries which have to use the go_library template and not just a
usual GN group(). A forwarding target for a go library is just another
go_library() with a (hopefully) unique name and a dependency on the
actual go_library(). The Go build rules aggregate dependencies through
library targets, so this ends up depending on the the real library and
whatever it depends on.

This requires restructuring the BUILD.gn generation logic slightly to
properly import the .gni file containing the go_library template
definition, since imports have to be unique and should be at the top of
the file.

Change-Id: Ibd96a0f510bd28681f47563b38eb547656883c18