[netstack3] Remove #![allow(unused)] from Eventloop mod
We've seen a few test failures where futures weren't awaited, and the
compiler didn't warn because we are allowing unused. This change removes
the mod-wide rule and fixes all of the new compiler warnings: unused
imports, things that are mut but shouldn't be, and unused return values
and variables.

Bug: related to 35497
Test: Functionality is not changed: fx run-test netstack3_tests runs all
existing unit tests and passes.

Change-Id: If91d1a23dca2daae48df66c7ee90d454238a00ec