[syscalls][kazoo] Implement the kernel-header backend
The biggest change here is that I realized some of the fidl was wrong,
in that it was using an *input* parameter like:

  mutable_vector_something data

where it should have been using an *output* like:

  vector_something data

This wasn't a problem in the previous backends because they ended up
being the same thing (just a non-const pointer), but in the kernel
header, we need to distinguish between user_out_ptr<> (the output
vector) and user_inout_ptr<> (a mutable input vector).

This involved another munging of syscall_library.cc's converter from
fidl-style to kernel-style. I think it's maybe a bit more
straightforward than before, but still a bit complex unfortunately.

Bug: 33668
Change-Id: I32f59a18353040c60eecbd27b5d1c4ee0de618a4