[debugger] Add Status IPC message.
This permits the client to query what is the current status of the debug
agent. For now the only query is to know what are the attached
processes, but this message can be extended in the future to support
other kind of queries.

This will permit the client to "formally" attach to those processes in
the case of a restarting session.

TEST=Add unit test.
     fx run-test debug_agent_tests

Change-Id: I269cf8570bcdfb8ae69e7ad5942bccc4846624a9