[audio_core] Replace dispatcher-pool with libasync.
This replaces usages of dispatcher-pool for managing mixing threads
with async::Loops. This change attempts to minimize the risk associated
with this change by using libasync in a way as congruent to
dispatcher-pool as possible, including:

  > Importing the Token definitions from dispatcher-pool to allow us to
    retain the thread annotations.
  > Providing a 'WakeupEvent' implemented on top of libasync.

Some aspects did require some changes due to semantic mismatches of how
the two libraries operate:

  > Some libasync primitives are not thread safe and can only be
    accessed from the dispatch thread they're bound to
    (ex: async::Wait/async::Task). To do safe initialization and
    shutdown we now need to run some operations on the mixer threads
    themselves that we could previously do on the main thread.

In order to provide some flexibility on how we allocate threads to
mixes, we also add an abstraction for acquiring the async_dispatcher_t*
for mixing in the ThreadingModel class. We initially support a few
basic threading models:

  > Mix everything on the main/fidl thread.
  > Create one thread for mixing that is distinct from the fidl thread.
  > Allocate a thread/async::Loop for every mix.

When compared to dispatcher-pool, the 'thread-per-mix' strategy is
most similar (dispatcher-pool maxes out at num_cpus, so on astro with
a representative workload we'll have 5 threads with libasync vs 4
threads with dispatcher-pool). The intention of the additional
threading models is to enable us to iterate on this to make an informed
decision on what's an optimal strategy here.

MULTIPLY: `[
  {
    "target": "//src/media/audio/audio_core/test:audio_fidl_tests",
    "total_runs": 10
  }, {
    "target": "//src/media/audio/audio_core/test:audio_admin_tests",
    "total_runs": 10
  }, {
    "target": "//src/media/audio/audio_core/test:audio_device_service_tests",
    "total_runs": 10
  }, {
    "target": "//src/media/audio/audio_core/test:audio_pipeline_tests",
    "total_runs": 10
  }, {
    "target": "//src/media/audio/audio_core:audio_core_unittests",
    "total_runs": 10
  }
]`

Test: fx run-test audio_core_unittest
        > Added new unit tests for WakeupEvent and ThreadingModel.
      fx run-test audio_device_tests
        > These will exercise the device add/remove path
      fx run-test audio_admin_tests
      - Manual sanity test on device
      - Used tracing to validate threads are properly released in the
        kThreadPerMix mode.
Change-Id: I8cebdc0e381abcff5f9fab4408f9cc0276b404b5