[debugger] Bring index implementations closer.
One more step in bringing the new index implementat (Index2) closer to
the old (Index).

This removes Index.FindPrefix which was unused, and implements
Index2.FindExact.

Fixes lots of bugs in the new indexer which the new unit test for
FindExact revealed.

Brings data member indexing up-to-date with recent changes in the normal
index to handle const values.

Change-Id: I328d34b4b42cbd55f1af17f6d0ba9ae63778cf0e