[bluetooth][snoop] Change verbosity arg to switch.
This fixes a bug where verbosity was an option that required the program
to be provided with a value rather than a flag. Now it's a flag where
the configured value is the number of occurrences of the flag.

Added tests to catch regressions and test that providing no arguments
produces the expected default values.

TEST=fx run-test bt-snoop-tests
BUG: 141132594

Change-Id: I04e291cf5a53b0103002af86e7e9b74f5d43936c