[ledger] Send obfuscated commit ids to the cloud.
Transmitting commit ids in the clear makes a potential attack where the
attacker tries to check the presence of a particular entry in the cloud
easier, since entry id generation depends on parent ids for merge
commits. This change makes the switch to using an obfuscated version of
the commit id when communicating with the cloud.

BUG:35364

Change-Id: I2cdffa80fdd9ee52374258bf74e7642e365436dd