[Settings] Add flush option for DeviceStorage.
This changelist adds a flag for write that indicates whether flush is
required for the operation to succeed. This will cause the operation to
block on the flush return. As a consequence, the write operation is now
async. The addition of flush rather than simply committing allows
controllers to now guarantee their changes are propagated to disk.

Fixes: fxb/36392
Test: fx run-test setui_service_tests -- --test
Testing strategy: An additional test has been added that triggers flush.
It is the only test that checks for flush as the subsequent action after
set from a write. Logging has also been added to the in memory proxies
so that flush commands can be verified from other tests. Existing tests
do explicitly check for commit, so those verify that flush is not called.

Change-Id: I05a4f2931bf6f2703d7196579584200814aa6d31