[fidl] Fix failing tests when auto-var-init is enabled
This CL addresses a few tests that fail when we use
-fauto-var-init=pattern with Clang. A number of tests attempt to
zero-initialize structs along the lines of:

```
recursion_message_layout message = {};
```

Deep into the heirarchy of these structs are unions which we intend to
zero-initialize. In the event where the first member of any of these
unions is not its largest member, only the bits pertaining to that first
member will be zero-initialized.

BUG: 3261
Change-Id: Id54bc434b61ba373fc430c8a48be6d83a876fdca