[debugger] Add a walker implementation for new index.
Adds an IndexWalker2 which walks the new Index implementation (Index2).
This is currently unused in production as we fill out the new index
implementation.

The new walker is more complex because there could be multiple paths to
a thing of a given name. This complexity and the solution is discussed
in the new header.

Change-Id: I23f87057f39767e91ed3159e03d69106bd1c39ea