[inspect] Fix bug with decrementing refcounts.
Deleting a Node before deleting its children should be possible, but
when the node has multiple children there is an error causing the node's
refcount to be reduced to 0 and deleted when one child is deleted.

This was caught by an assertion in debug mode, but in release mode it
could cause data corruption.

Change-Id: I0e66377222aa0fe23294467a658dcb704df7643b