[debugger] Handle anonymouns namespaces in new index.
In the old index (the currently used implementation) just omitted
anonymous namespaces and types and indexed the names without them. This
makes it impossible to explicitly search for things in anonymous
namespaces so the new index includes this information.

But having the anonymous namespaces interferes with name searching. This
patch makes the new index walker implicitly go into anonymous
namespaces.

Change-Id: Ib43f56dba0ecbe5b8c78019c4c857daba09b228c