[netstack3] Gate enable_interface on logical AND admin + phy
To enable an interface, the netstack should logical AND admin and phy
statuses. To disable an interface, the netstack should OR it (allow any
call to disable). Now the netstack caches phy state in the bindings and
has infrastructure to modify cached state and query before enabling.

Bug: 21135 #comment
Testing: Added unit test to ensure device state is handled correctly in
the bindings and core.
`fx run-test netstack3_tests`

Change-Id: I451c68c0a85a4acbfc790eed4d2f9ecac5e98628