[debugger] Don't index anonymous types.
Previously the new index would create index entries for anonymous
structs and enums would would cause problems because they would
implicitly match the containing type when looking up names. This change
makes those types have nodes in the index, but only to hold the
children, not containing any DIEs.

There was already an anonymous enum in the test symbol file that
exibited problematic behavior, but we didn't encounter it because the
index dump uses the checked-in file which was made before the anonymous
enum was added.

This updates the checked-in symbol file. The instructions are moved from
the BUILD file to a README in the same directory since a separate target
is no longer needed.

The index dump and some constants are updated with the new expectations.

Change-Id: I462bb6ca97d29af602d7f5198c40a59754155cb0