[build] Update //tools:infratools
Updates the toolchains that tools:infratools are built with.
On mac, just the host toolchain; on linux, the linux arm64/x64
toolchains. This is to preserve the invariant that any tool in
tool_paths.json has a unique (name, os, cpu) signature - and that arm
tools are available on linux hosts.

Change-Id: I68e21275824107caf0fb5a1fb70079e2d69e5670