[minfs] reduce backup superblock write frequency.
This change reduces the amount of writes of backup superblock.
namely PersistAllocate and PersistRelease since these values can be recreated
using fsck repair.

It does not change the backup update for PersistentStorage::Extend since this
event is less frequent than PersistAllocate and PersistRelease.

testing: Added an integration test TestAllocCountWriteFrequency, which checks the
         alloc_counts after mount for superblock and backup, expects them to be equal,
         then explicitly does a mkdir to force inode and data block allocation.
         checks superblock and backup alloc_counts again and expects them to be unequal this time.
         Finally unmounting the filesystem should make these counts consistent again.

Change-Id: Id9f110ca78f8f412c86ae4c44cf280a0caa38349