[modular] inspect is_deleted propery added to mod inspect tree
to view the inspect tree:

iquery --find . | grep  "sessionmgr"
cd hub/c/sessionmgr.cmx/{session id numbers}/out/inspect
iquery --recursive root.inspect

The mod node and its properties should disappear from the inspect tree
when the mod is deleted. The property will only be important in the
case that the deletion fails and the mod is not removed from the
RunningModInfo struct.

Change-Id: I07e390b42f345065ff5e52b45063fb162a03a1b9