[zircon] Fix failing tests when using -fauto-var-init=pattern
This addresses tests /boot/test/sys/optee-message-test and
/boot/test/sys/pci-driver-test when using automatic variable
initialization.

The issue is that structs that are zer-initializes are not entirely
filled with zeros. This can happen if the struct contains a nested union
whose first member is not that union's largest. This can be addressed by
memset-ing the entire struct.

BUG: 3261
Change-Id: I0784ea7dc1afedaf876120eb184e651404d730f9