[connectivity][ppp] create serial PPP driver
This change creates a Point-to-Point Protocol driver over a serial
link layer. The driver exposes the Generic Frame Device FIDL interface.
The driver implements the synchronous byte-stuffing framing method
described in RFC-1662.

In future work, a service will be created which will handle the details
of PPP handshakes and session management, and this driver will integrate
with the service to fully expose the Generic Frame Device interface.

Test: `fx run-test ppp-tests`
Change-Id: I0d31061ca4c08338a5570a1195cb4db8d5789634