[scenic] have each session own its FIDL binding
Each session now owns its own FIDL binding of the Session protocol, as
opposed to Scenic owning all in a BindingSet.

This change is necessary to support FIDL events that are incoming, such
as OnFramePresented().

Tests: existing unit tests, and additional scenic_unittest for multiple
out of order client deletions

Change-Id: Ifa37acf8b440e6272350cb7caaa311217e35fa40