[hwinfo][sdk] Add BoardInfo Protocol and update ProductInfo Protocol in Hwinfo.fidl interface
The Product specific configurations need to be surfaced in a common
API interface. Hence, adding the following Product specific fields
into ProductInfo Protocol.
- product_name
- product_model
- manufacturer

Additionally, Board information is common across a group of devices
which mandates the need to introduce the BoardInfo protocol with the
following to fields:
- board_name
- board_revision

Bug: b/136513729
Design: go/fuchsia-hwinfo-component

Change-Id: Ib335cea3c1b1f1f45e7c505f4960b182baeca174